implement slack signature verification (#40)
